# cmd/docs/swagger.yaml
basePath: /api/v1
definitions:
  attendance.Attendance:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      date:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/attendance.AttendanceState'
      worker:
        $ref: '#/definitions/workers.Worker'
      worker_id:
        type: integer
    type: object
  attendance.AttendanceState:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - CheckedIn
    - CheckedOut
    - Expired
...
  common.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  departments.Department:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dtos.CreateAttendanceDto:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      date:
        type: string
      worker_id:
        type: integer
    required:
    - check_in
    - check_out
    - date
    - worker_id
    type: object
  dtos.CreateDepartmentDto:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
  dtos.CreatePayrollDto:
    properties:
      base_salary:
        type: number
      bonus:
        minimum: 0
        type: number
      deductions:
        type: number
      month:
        type: string
      worker_id:
        type: integer
    required:
    - base_salary
    - month
    - worker_id
    type: object
...
  dtos.UpdateAttendanceDto:
    properties:
      check_out:
        type: string
      id:
        type: integer
    required:
    - id
    type: object
  dtos.UpdateDepartmentDto:
    properties:
      description:
        maxLength: 255
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - id
    - name
    type: object
...
  pagination.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  workers.Worker:
    properties:
      department:
        $ref: '#/definitions/departments.Department'
      department_id:
        type: integer
      email:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple ERP API
  title: User API
  version: "1.0"
paths:
...
  /attendances:
    get:
      description: Get attendances with optional pagination query parameters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.PaginationResponse'
      security:
      - BearerAuth: []
      summary: List attendances (paginated)
      tags:
      - attendances
  /attendances/{id}:
    get:
      description: Get an attendance record by its ID
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/attendance.Attendance'
      security:
      - BearerAuth: []
      summary: Get an attendance record by ID
      tags:
      - attendances
  /attendances/checkin:
    post:
      consumes:
      - application/json
      description: Create a new attendance record for a worker
      parameters:
      - description: Attendance payload
        in: body
        name: attendance
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateAttendanceDto'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create an attendance record
      tags:
      - attendances
  /attendances/checkout:
    put:
      consumes:
      - application/json
      description: Update the check-out time of an existing attendance record
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated attendance payload
        in: body
        name: attendance
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateAttendanceDto'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update an attendance record
      tags:
      - attendances
  /attendances/worker/{worker_id}:
    get:
      description: Get all attendance records for a specific worker
      parameters:
      - description: Worker ID
        in: path
        name: worker_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/attendance.Attendance'
            type: array
      security:
      - BearerAuth: []
      summary: Get attendance records by worker ID
      tags:
      - attendances
...
  /departments:
    get:
      description: Get departments with optional pagination query parameters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.PaginationResponse'
      security:
      - BearerAuth: []
      summary: List departments (paginated)
      tags:
      - departments
    post:
      consumes:
      - application/json
      description: Create a new department
      parameters:
      - description: Department payload
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateDepartmentDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateDepartmentDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIError'
      security:
      - BearerAuth: []
      summary: Create a department
      tags:
      - departments
    put:
      consumes:
      - application/json
      description: Update an existing department
      parameters:
      - description: Department payload
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateDepartmentDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.UpdateDepartmentDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIError'
      security:
      - BearerAuth: []
      summary: Update a department
      tags:
      - departments
  /departments/{id}:
    delete:
      description: Delete a department by ID
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a department
      tags:
      - departments
    get:
      description: Get a department by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/departments.Department'
      security:
      - BearerAuth: []
      summary: Get a department
      tags:
      - departments
  /payroll/:workerId:
    get:
      description: Get payrolls by worker ID
      parameters:
      - description: Worker ID
        in: path
        name: workerId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get payrolls for a worker
      tags:
      - payrolls
  /payroll/calculate:
    post:
      consumes:
      - application/json
      description: Calculate net salary for a payroll payload
      parameters:
      - description: Payroll payload
        in: body
        name: payroll
        required: true
        schema:
          $ref: '#/definitions/dtos.CreatePayrollDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIError'
      security:
      - BearerAuth: []
      summary: Calculate payroll
      tags:
      - payrolls
  /payrolls:
    get:
      description: Get payrolls with optional pagination query parameters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.PaginationResponse'
      security:
      - BearerAuth: []
      summary: List payrolls (paginated)
      tags:
      - payrolls
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer <JWT_TOKEN>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"